requirements.txt
streamlit>=1.36.0
openai>=1.37.0
python-dotenv>=1.0.1
pydantic>=2.7.4